I have implemented a basis for the algorithms
WILL PROBABLY CONTAIN LITTLE BUGS

- Implemented replay buffer (experience replay)
- Implemented GNN
- Implemented Deep Q-Learning agent

I took the code from my book: https://lars-quaedvlieg.github.io/RL-Playground/content/deep-rl/dqn.html

and adapted it to this problem, but we still need to test whether it works (we can only do that once we main loop is made). Additionally, we still need to add some things like action masking for lane changes that are not allowed

I also changed the name of the RLAgent to DQNAgent, changed the name of the "getDecision" method to "choose_action", and changed the "fetchVehicleFeatures" method to "store_transition" with a new method signature, since we need that for the RL agent. We  can modify the main loop to accomodate this once we have the changes from Erik.